Add fullsupport Accounts, Products, and Orders API
Refactor client design to DRY code a bit more when generating HTTP
requests.

Change the package name from cb to coinbase as it adds clarity and
a user can alias to cb if they desire a abbreviated package name.

Make use of go-querystring for URL query params to sipmplify the
process of turning optional config structs to a query string.