Comply with GNU licensing guidelines
Add a copyright notice at the top of each file.

Add a license notice at the top of each file to remove uncertainty
about whether the license applies to only a portion of the program. As
advocated by GNU https://www.gnu.org/licenses/gpl-howto.html in their
documentation.

Add a source link that leads users to an archive of the code inside
the doc.go file. This is to comply with GNU AGPL.