Add the list portfolios API
Add the PortfoliosService to the client to enable further integration
of the rest of the portfolio APIs.

Add list all portfolios API to retrieve a list of all portfolios of a
user.

Add a new get helper funcion on the client to remove a lot of duplicate
code when building and handling the HTTP request. This should DRY the
code quite a bit once the rest gets refactored.

Remove unnecessary newline in orders_list.go for linting.