Add support for Cloud API Trading keys
The client now supports Coinbase Cloud API Trading keys as an
authentication scheme. Provide a constructor NewWithCloud() to handle
token generation and handle interaction with HTTP requests.

Refactor the client from using default legacy API authentication to a
more generic method. Expose a public interface to allow the consumer
to specify a custom implementation. This prevents me from having to
create a handler for the OAuth scheme since it would never be a use
case for me. Also better practice :)

Update README with info on new authentication schemes and how to use
them.